+++ models/ShopFuser.php
@@ -61,7 +61,7 @@ use yii\helpers\ArrayHelper;
  * @property CmsContentElement $store
  *
  */
-class ShopFuser extends Core
+class ShopFuser extends Core implements \JsonSerializable
 {
     /**
      * @inheritdoc
@@ -148,6 +148,25 @@ class ShopFuser extends Core
             'countShopBaskets',
             'shopBaskets',
         ];
+    }
+
+    /**
+     *
+     * Массив для json ответа, используется при обновлении корзины, добавлении позиций и т.д.
+     *
+     * @return array
+     */
+    public function jsonSerialize()
+    {
+        return ArrayHelper::merge($this->toArray([], $this->extraFields()), [
+            'money'                     => ArrayHelper::merge($this->money->jsonSerialize(), ['convertAndFormat' => \Yii::$app->money->convertAndFormat($this->money)]),
+            'moneyDelivery'             => ArrayHelper::merge($this->moneyDelivery->jsonSerialize(), ['convertAndFormat' => \Yii::$app->money->convertAndFormat($this->moneyDelivery)]),
+            'moneyDiscount'             => ArrayHelper::merge($this->moneyDiscount->jsonSerialize(), ['convertAndFormat' => \Yii::$app->money->convertAndFormat($this->moneyDiscount)]),
+            'moneyOriginal'             => ArrayHelper::merge($this->moneyOriginal->jsonSerialize(), ['convertAndFormat' => \Yii::$app->money->convertAndFormat($this->moneyOriginal)]),
+            'moneyVat'                  => ArrayHelper::merge($this->moneyVat->jsonSerialize(), [
+                'convertAndFormat' => \Yii::$app->money->convertAndFormat($this->moneyVat)
+            ]),
+        ]);
     }
 
     /**